Brain-Tasks-App/lambda-deploy/requirements.txt
boto3==1.26.137
kubernetes==26.1.0
PyYAML==6.0
